# docs/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Cloudflare Backend Boilerplate API",
		"version": "0.1.0",
		"description": "API-only Cloudflare Worker (Hono) for mobile clients. Generated from src/routes/* and src/lib/schemas.ts — see scripts/generate-openapi.ts."
	},
	"servers": [
		{
			"url": "http://127.0.0.1:8787",
			"description": "Local dev"
		}
	],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Better Auth JWT access token (15m). Also accepts an opaque session/API token on some routes — see docs/README."
			}
		}
	},
...
	"paths": {
...
		"/api/v1/auth/register": {
			"post": {
				"tags": [
					"Auth"
				],
				"summary": "Register with email + password",
				"responses": {
					"200": {
						"description": "User created; tokens issued when email/password sign-up completes immediately",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string",
										"format": "email"
									},
									"password": {
										"type": "string",
										"minLength": 1
									},
									"name": {
										"type": "string",
										"minLength": 1
									}
								},
								"required": [
									"email",
									"password",
									"name"
								],
								"additionalProperties": false
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/login": {
			"post": {
				"tags": [
					"Auth"
				],
				"summary": "Log in with email + password",
				"responses": {
					"200": {
						"description": "User + access/refresh tokens",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string",
										"format": "email"
									},
									"password": {
										"type": "string",
										"minLength": 1
									}
								},
								"required": [
									"email",
									"password"
								],
								"additionalProperties": false
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/logout": {
			"post": {
				"tags": [
					"Auth"
				],
				"summary": "Revoke the current session",
				"responses": {
					"200": {
						"description": "Session revoked",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					},
					"401": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				]
			}
		},
		"/api/v1/auth/refresh": {
			"post": {
				"tags": [
					"Auth"
				],
				"summary": "Exchange a refresh token for a new access/refresh pair",
				"responses": {
					"200": {
						"description": "New access/refresh tokens",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"refreshToken": {
										"type": "string",
										"minLength": 1
									}
								},
								"required": [
									"refreshToken"
								],
								"additionalProperties": false
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/verify-email": {
			"post": {
				"tags": [
					"Auth"
				],
				"summary": "Verify an email address with a token",
				"responses": {
					"200": {
						"description": "Verification status + user",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"token": {
										"type": "string",
										"minLength": 1
									}
								},
								"required": [
									"token"
								],
								"additionalProperties": false
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/resend-verification": {
			"post": {
				"tags": [
					"Auth"
				],
				"summary": "Resend the verification email",
				"responses": {
					"200": {
						"description": "Email sent when the account exists",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string",
										"format": "email"
									}
								},
								"required": [
									"email"
								],
								"additionalProperties": false
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/forgot-password": {
			"post": {
				"tags": [
					"Auth"
				],
				"summary": "Request a password reset email",
				"responses": {
					"200": {
						"description": "Reset email sent when the account exists",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string",
										"format": "email"
									}
								},
								"required": [
									"email"
								],
								"additionalProperties": false
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/reset-password": {
			"post": {
				"tags": [
					"Auth"
				],
				"summary": "Reset a password with a reset token",
				"responses": {
					"200": {
						"description": "Password reset status",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"token": {
										"type": "string",
										"minLength": 1
									},
									"newPassword": {
										"type": "string",
										"minLength": 1
									}
								},
								"required": [
									"token",
									"newPassword"
								],
								"additionalProperties": false
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/google": {
			"get": {
				"tags": [
					"Auth"
				],
				"summary": "Start the Google OAuth browser flow",
				"responses": {
					"200": {
						"description": "Authorization URL to redirect the browser to",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": [
					"Auth"
				],
				"summary": "Sign in with Google (native ID token)",
				"responses": {
					"200": {
						"description": "User + access/refresh tokens",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"idToken": {
										"type": "string",
										"minLength": 1
									},
									"nonce": {
										"type": "string"
									},
									"accessToken": {
										"type": "string"
									}
								},
								"required": [
									"idToken"
								],
								"additionalProperties": false
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/apple": {
			"post": {
				"tags": [
					"Auth"
				],
				"summary": "Sign in with Apple (native identity token)",
				"responses": {
					"200": {
						"description": "User + access/refresh tokens",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"identityToken": {
										"type": "string",
										"minLength": 1
									},
									"authorizationCode": {
										"type": "string"
									},
									"fullName": {
										"type": "object",
										"properties": {
											"givenName": {
												"type": "string"
											},
											"familyName": {
												"type": "string"
											}
										},
										"additionalProperties": false
									}
								},
								"required": [
									"identityToken"
								],
								"additionalProperties": false
							}
						}
					}
				}
			}
		},
		"/api/v1/auth/change-password": {
			"post": {
				"tags": [
					"Auth"
				],
				"summary": "Change the signed-in user's password",
				"responses": {
					"200": {
						"description": "Password changed",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"data": {}
									}
								}
							}
						}
					},
					"400": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					},
					"401": {
						"description": "Structured error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"error": {
											"type": "object",
											"properties": {
												"code": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"details": {}
											},
											"required": [
												"code",
												"message"
											]
										}
									},
									"required": [
										"error"
									]
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"currentPassword": {
										"type": "string",
										"minLength": 1
									},
									"newPassword": {
										"type": "string",
										"minLength": 8
									}
								},
								"required": [
									"currentPassword",
									"newPassword"
								],
								"additionalProperties": false
							}
						}
					}
				}
			}
		},
...
	}
}